Add UI state machine builder
Add the cypress.state-machine namespace, which gives you tools to define
and validate UI state machines. UI state machines are state machines
where the transitions are triggered on DOM events and cause the
application state to be updated. (The application state update functions
will have to take the current app state and the event that caused the
transition to fire and return a new app state, but since these state
machines aren't used for anything yet that constraint isn't enforced.)

Both the representation implementation and the validation functions have
tests.