Add SM states fn, refactor transitions-from
Add the `states` function to the state machine ns, which returns
a sequence of all the distinct states in the state machine that you give
it.

Change `transitions-from` to take a full state machine representation
and the state you're interested in, rather than the sequence of
transitions, since this is easier for external consumers.

Rename the old `transitions-from` function to be the new private
function `transitions-from*`, and change the helper functions for
`connected?` to use the new name.