Add unroll-machine fn to core.
Add a function to transform a state machine representation as generated
by the state-machine ns into the form that will be used inside the event
loop. This form is a doubly-nested map keyed by originating state and
event, whose values have the update function to apply to the app state
and and the state to transition to. This makes looking up what to do
& what state to switch to as new events come in a matter of doing
a simple `(get-in sm [current-state new-event-type])`.