Implement app state updating in event loop.
Add an app state atom argument to the core init fn and the event
processor constructor function. The value in the atom at the time it's
passed to init is captured as the initial app state. When transitions
are triggered, the event loop iteratively applies the update functions
from the UI state machine to the state value to obtain the new current
value, and the atom is updated to hold the current value.

Add placeholder names to the unused arguments of the state machine ns's
default identity-update state transition function, to make the interface
of update functions more clear.