Support more events, only listen for what's used
Add in the rest of the DOM & touch events that I could think of / find
in the MDN documentation.

In order to avoid registering a bunch of useless listeners that don't
actually trigger transitions in the application's UI state machine, only
register handlers for events that are used in some transition of the
state machine.