Allow dispatched transitions in state machines
Allow the `to` argument in `add-transition` to be a "dispatch" function
that returns the state to transition to, instead of having to be fixed.
This does have the consequence that the state machine's graph can no
longer be statically checked for connectedness (since the behavior of
the dispatch function is unknown), so add a `dispatched-transition?`
predicate on state machines and use it to avoid the connectedness check
for state machines with any dispatched transitions.