Support dispatch transitions in event loop
When a transition is triggered in the event loop, call the new
`transition-target` fn on the `to` entry. `transition-target` abstracts
away caring about whether `to` is a keyword like before or a dispatch
function; it returns the state the transition goes to either way.

With this addition, dispatch functions are now fully supported.
Hopefully this makes the first version of Cypress functionally complete.