Accept core.async channels as custom event sources
Add a new optional argument to the `init!` function which is
a map of event kind keywords to core.async channels that produce custom
events of that kind that are processed the UI state machine. The
custom events are completely opaque to cypress; the user only needs to
ensure that their custom event channels and the transition functions
they define for the UI state machine agree on the events' structure.