Allow adding transitions with custom triggers.
A transition has a 'custom trigger' when it has a function for its 'on'
parameter (instead of a DOM event keyword). The function takes any event
that fires in its 'from' state and returns truthy when the transition
should trigger. These functions are called 'recognition functions',
though that is mostly an implementation term.

Also add tests ensuring that these custom-triggered transitions can be
added to state machines.

Note that while they are allowed in state machines, the core event loop
does not yet support them, so they won't do anything.