Support custom triggers in the event loop
When no DOM trigger match is found in a state, pass the event to all the
recognition functions of custom-triggered transitions that start in that
state to see if any of them fire.

Doing this efficiently in the event loop is supported by sticking all
the outward transitions with custom triggers in a special key in each
state's map in an unrolled state machine.