Break out README's guided example to its own doc
It's really a bit too involved for a README, and crowds the rest out.
Plus I'll be adding more documentation explaining the other examples.

Right now there's a FIXME in the doc to come back to, concerning the gap
between what's in the doc and the cljsbuild & figwheel dependencies
& configuration you'd need to add to the project.clj in order to
actually get it to work if you're following the doc instead of just
reading it.