Add third guide, on custom events.
This accompanies the 'calculator' example.

In the course of writing this I realized that the motivation for passing
a flat collection of custom event channels and then using custom trigger
functions to disambiguate them is not strong, and it's better for both
consumers and Cypress itself to require always passing a map of event
category keyword to a channel that produces those events. If consumers
really need to create categories out of a single source (e.g.
a websocket that sends different categories of events back to the app),
they can always write their own go block that sorts the events into
channels that are then fed into Cypress.

To that end, refactor the calculator example a bit to use the map
form of custom events, and only cover that in the guide. (I still
need to go and remove the support for custom triggers from Cypress and
scrub all references to those from the docstrings).

Make a few other name changes and simplifications in the calculator, in
the interest of clarity.